petpal/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
Pillow